add (disabled) tests to look for object leakage
this uses a tool to find the set of all objects reachable from some starting
point, then compares those sets for intersection

The tests are disabled because of known failures: the function-repair code in
accessors.js puts __defineGetter__/etc properties on 'Function', but those
properties inherit from the primal Realm's Object, and thus leak access.

The fix will be to execute those repair functions *as a string* inside each
new RootRealm, as a shim.