reject dynamic import expressions in eval() and new Function()
We originally considered doing this as part of a shim that would be injected
into each new RootRealm (wrapping the eval/Function globals), but there are
too many cases to do this easily. Now we just apply it in all RootRealms when
building the safe evaluator.

We use a conservative regexp to spot 'import' keywords. This will reject
several safe forms, including commented out import() calls (making them
taboo: even talking about a dynamic import will cause the script to be
rejected). If/when a full parser is available, we should be able to make this
safe and accurate.

We do not look for `import.`, since 'import.meta' is not allowed in script
contexts, so we don't need to guard against it.

The security of this regexp depends it matching the behavior of the parser,
specifically they must agree on what qualifies as whitespace. There is a test
case that exhaustively checks all unicode code points for agreement, but it
takes too long to every time (35s on my laptop), so it's disabled. We should
run it at least once on any new platform to make sure there are no surprises.