fix sandbox bypass in Function constructor
The `Function(arg1, arg2, body)` constructor works by building a function
definition out of its parameters and then evaluating the generated string. It
is supposed to inspect the parameters to prevent access to global variables,
but Matt Austin found a clever bypass involving back-to-back template
literals.

This allowed execution of arbitrary code in the "safe" global environment,
giving `Function()` authority comparable to our safe `eval()` facility. This
wouldn't give a child Realm any additional power, but if the Realm was using
`Function()` to test that a given untrusted string was a well-formed function
body without actually executing it, this bug would give the source of that
string a surprising temporal power. The attacker's expression would be
executed immediately, whereas function constructors do not execute the
generated function at all (they merely produce a callable value which could
be invoked later).

The attacker's expression can also access `this`, which function bodies
cannot normally reach, but since functions can simply do `eval("this")`, this
is not a net increase in power.

Our fix is to reject all but the simplest of function parameters. We removed
default parameters, destructuring assignment, and the "rest"
parameter (`foo(arg1, ...rest)`), and only allow simple ASCII-based
identifier names.

refs #193