shim: build TamedFunction as a strict function, not sloppy
The previous use of 'Function' actually created a "sloppy-mode"
function (whereas using 'function' creates a "strict-mode" function). This
was a leak: we should not be allowing sloppy-mode functions to be accessible
within the new Realm. The sloppy-mode function had '.arguments' and '.caller'
properties, which confused the SES code that removes non-whitelisted
properties.