Normative: Remove evaluate method
Rationale: To get to a minimal Realms MVP, the functionality can be accessed
by `realm.globalThis.eval` (stash off to the side soon after creating realm to
use reliably).

Omitting this method for now would maximize the flexibility of the continued
development of the Compartment/Evaluator API; that proposal could add an
analogous method to Realm.prototype as part of the same proposal.

Closes #226
Alternative to #224