Editorial: Call GetWrappedValue with completion record's [[Value]]
There's a confusion of completion records and language values in two
uses of the GetWrappedValue AO - quoting ECMA-262:

    Within this specification, the notation “Type(x)” is used as
    shorthand for “the type of x” where “type” refers to the ECMAScript
    language and specification types defined in this clause. [...]

An Object is a language type, a completion record is a specification
type.
This means that the "If Type(value) is Object" path in GetWrappedValue
would never be taken for these calls, even if the completion cecord's
[[Value]] field is an Object. Correct this by passing result.[[Value]]
instead of result (the completion record) itself.