Update ecmarkup and biblio, structured headers, minor spec language issues (#389)
* Fix punctuation/typos

* Update ecmarkup

Update from v15 to v18. Fix new linter issues: no need to include script
and stylesheets manually, and fix partially-multiline if-clauses.

* Rebase CopyNameAndLength on ECMA-262 Function.prototype.bind

The language has changed slightly in ECMA-262's Function.prototype.bind
definition. Update the deleted steps and CopyNameAndLength to match.

* Rebase PerformShadowRealmEval on ECMA-262 eval()

Update some of the language in PerformShadowRealmEval to match updated
language in ECMA-262 eval().

* Update 'subclassable' language

ECMA-262 removed the "is designed to be subclassable" language, preferring
to communicate the fact that a constructor can be extended, without making
any value judgement as to whether it was designed to do so.

* Update Type() language to "is a"

ECMA-262 now uses "is an Object", "is not a String", etc., instead of the
Type() language.

* Fix presence handling of optional arguments in CopyNameAndLength

Absent argument is not the same as undefined. According to Shu, there is
no syntax for perfect forwarding in ECMA-262 language.

* Break out anonymous built-in functions into own sections

Following the example of ECMA-262 such as in Promise Reject Functions, the
anonymous built-in functions should be placed in their own subsections,
and their steps referred to using emu-xref.

* Update "performs the following steps when called" language

ECMA-262 writes the preamble to function algorithm steps like this now.

* Update structured headers

The structured headers should get return types, and we should use a
structured header for the host-defined operation.

---------

Co-authored-by: Caridy Patiño <caridy@gmail.com>